requirements.txt
adal==1.2.4
azure-cognitiveservices-search-websearch==2.0.0
azure-common==1.1.25
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
emoji==0.6.0
html5lib==1.1
idna==2.10
isodate==0.6.0
lxml==4.5.2
msrest==0.6.18
msrestazure==0.6.4
multitasking==0.0.9
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.7
urllib3==1.25.10
webencodings==0.5.1
yfinance==0.1.54
